jaegar/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.15
Django==5.0.7
django-prometheus==2.3.1
djangorestframework==3.15.2
flower==2.0.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
gunicorn==23.0.0
humanize==4.10.0
idna==3.10
importlib_metadata==8.5.0
kombu==5.3.7
mysqlclient==2.2.4
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-django==0.50b0
opentelemetry-instrumentation-wsgi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.29.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.0
zipp==3.21.0
